Create Student_GPA_Avg.c
This simple program lets the user input up to 30 students GPA's. The user will make their selection and pending will be the result. For each entry they should see "for student number" and the number will be displayed for which student they are on. Once the user is ready and they press 2, the program will find the average of GPA's for said students.